feat: useFetchCities 타입 지정 / city 타입 빼기 / 쿼리키 파일에 분리
