+++ src/types/city.ts
@@ -1,0 +1,9 @@
+export interface CitiesType {
+  cityId: number;
+  cityName: string;
+  countryName: string;
+}
+
+export interface CitiesResponse {
+  cities: CitiesType[];
+}
