+++ src/utils/constants/api/queryKey.ts
@@ -1,3 +1,4 @@
 export const queryKey = {
   HOTELS_LIST: 'hotelsList',
+  CITIES: 'cities'
 } as const;
